homeworks/nn-reranking/requirements.txt
numpy
hydra-core
pandas
nltk
emoji
torch
torchmetrics
lightning
transformers
composer
torch_ema
wandb